float64->float32

+++ src/TextgenConnectionMap.ts
@@ -33,7 +33,7 @@ export type TextgenConnectionMap = {
         prompt: string;
       };
     };
-    response: Float64Array;
+    response: Float32Array;
     headers: {
       url?: string;
       authToken?: string;
